feat: refactor AccountDetails into smaller components (#55)
