+++ src/app/AccountDetails/AccountDetails.tsx
@@ -1,101 +1,16 @@
-import { parseScanTimestamp, parseNumberToCurrency } from 'src/app/utils/parseFuncs';
-import { renderStatusLabel } from 'src/app/utils/renderStatusLabel';
 import React, { useEffect, useState } from 'react';
 import { useParams } from 'react-router-dom';
-import {
-  PageSection,
-  PageSectionVariants,
-  Tabs,
-  Tab,
-  TabContent,
-  TabContentBody,
-  TabTitleText,
-  Title,
-  DescriptionList,
-  DescriptionListGroup,
-  DescriptionListTerm,
-  DescriptionListDescription,
-  Label,
-  Flex,
-  FlexItem,
-  Page,
-} from '@patternfly/react-core';
-import { Table, Tbody, Td, Th, Thead, Tr } from '@patternfly/react-table';
-import { getAccountByName, getAccountClusters } from '../services/api';
-import { Link } from 'react-router-dom';
-import { AccountData, Cluster } from '@app/types/types';
-import { LoadingSpinner } from '../components/common/LoadingSpinner';
-
-const columnNames = {
-  id: 'ID',
-  name: 'Name',
-  status: 'Status',
-  cloudProvider: 'Cloud Provider',
-  instanceCount: 'Instance Count',
-};
-
-const AggregateClustersPerAccount: React.FunctionComponent = () => {
-  const [data, setData] = useState<Cluster[] | []>([]);
-  const [loading, setLoading] = useState(true);
-  const { accountName } = useParams();
-
-  useEffect(() => {
-    const fetchData = async () => {
-      try {
-        console.log('Fetching data...');
-        const fetchedAccountClusters = await getAccountClusters(accountName);
-        console.log('Fetched Account data:', fetchedAccountClusters);
-        setData(fetchedAccountClusters);
-      } catch (error) {
-        console.error('Error fetching data:', error);
-      } finally {
-        setLoading(false);
-      }
-    };
-
-    fetchData();
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, []);
-
-  console.log('Rendered with data:', data);
-
-  return (
-    <React.Fragment>
-      {loading ? (
-        <LoadingSpinner />
-      ) : (
-        <Table aria-label="Simple table">
-          <Thead>
-            <Tr>
-              <Th>{columnNames.id}</Th>
-              <Th>{columnNames.name}</Th>
-              <Th>{columnNames.status}</Th>
-              <Th>{columnNames.cloudProvider}</Th>
-              <Th>{columnNames.instanceCount}</Th>
-            </Tr>
-          </Thead>
-          <Tbody>
-            {data.map((cluster) => (
-              <Tr key={cluster.id}>
-                <Td dataLabel={cluster.name}>
-                  <Link to={`/clusters/${cluster.id}`}>{cluster.id}</Link>
-                </Td>
-                <Td>{cluster.name}</Td>
-                <Td dataLabel={cluster.status}>{renderStatusLabel(cluster.status)}</Td>
-                <Td>{cluster.provider}</Td>
-                <Td>{cluster.instanceCount}</Td>
-              </Tr>
-            ))}
-          </Tbody>
-        </Table>
-      )}
-    </React.Fragment>
-  );
-};
+import { Page } from '@patternfly/react-core';
+import { getAccountByName } from '@app/services/api';
+import { AccountData } from '@app/types/types';
+import AccountsHeader from './components/AccountHeader';
+import AccountsTabs from './components/AccountTabs';
+import { AccountDetailsContent } from './components/AccountDetailsContent';
+import { debug } from '@app/utils/debugLogs';
+import { AccountClusters } from './components/AccountClusters';
 
 const AccountDetails: React.FunctionComponent = () => {
-  const { accountName } = useParams();
-  const [activeTabKey, setActiveTabKey] = React.useState(0);
+  const { accountName } = useParams() as { accountName: string };
   const [accountData, setAccountData] = useState<AccountData>({
     count: 0,
     accounts: [],
@@ -104,10 +19,10 @@ const AccountDetails: React.FunctionComponent = () => {
   useEffect(() => {
     const fetchData = async () => {
       try {
-        console.log('Fetching Account Clusters ', accountName);
+        debug('Fetching Account Clusters ', accountName);
         const fetchedAccountClusters = await getAccountByName(accountName);
         setAccountData(fetchedAccountClusters);
-        console.log('Fetched Account Clusters data:', accountData);
+        debug('Fetched Account Clusters data:', fetchedAccountClusters);
       } catch (error) {
         console.error('Error fetching data:', error);
       } finally {
@@ -116,99 +31,15 @@ const AccountDetails: React.FunctionComponent = () => {
     };
 
     fetchData();
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, []);
-
-  const handleTabClick = (event, tabIndex) => {
-    setActiveTabKey(tabIndex);
-  };
-
-  const detailsTabContent = (
-    <React.Fragment>
-      {loading ? (
-        <LoadingSpinner />
-      ) : (
-        <Flex direction={{ default: 'column' }}>
-          <FlexItem spacer={{ default: 'spacerLg' }}>
-            <Title headingLevel="h2" size="lg" className="pf-v5-u-mt-sm" id="open-tabs-example-tabs-list-details-title">
-              Account details
-            </Title>
-          </FlexItem>
-
-          <FlexItem>
-            <DescriptionList
-              columnModifier={{ lg: '2Col' }}
-              aria-labelledby="open-tabs-example-tabs-list-details-title"
-            >
-              <DescriptionListGroup>
-                <DescriptionListTerm>Name</DescriptionListTerm>
-                <DescriptionListDescription>{accountData.accounts[0].name}</DescriptionListDescription>
-                <DescriptionListTerm>Account ID</DescriptionListTerm>
-                <DescriptionListDescription>{accountData.accounts[0].id}</DescriptionListDescription>
-              </DescriptionListGroup>
-              <DescriptionListGroup>
-                <DescriptionListTerm>Cloud Provider</DescriptionListTerm>
-                <DescriptionListDescription>{accountData.accounts[0].provider}</DescriptionListDescription>
-                <DescriptionListTerm>Account Total Cost (Estimated)</DescriptionListTerm>
-                <DescriptionListDescription>
-                  {parseNumberToCurrency(accountData.accounts[0].totalCost)}
-                </DescriptionListDescription>
-              </DescriptionListGroup>
-              <DescriptionListGroup></DescriptionListGroup>
-              <DescriptionListGroup>
-                <DescriptionListTerm>Last scanned at</DescriptionListTerm>
-                <DescriptionListDescription>
-                  {parseScanTimestamp(accountData.accounts[0].lastScanTimestamp)}
-                </DescriptionListDescription>
-              </DescriptionListGroup>
-              <DescriptionListGroup></DescriptionListGroup>
-            </DescriptionList>
-          </FlexItem>
-        </Flex>
-      )}
-    </React.Fragment>
-  );
-
-  const clustersTabContent = (
-    <TabContentBody>
-      <AggregateClustersPerAccount />
-    </TabContentBody>
-  );
+  }, [accountName]);
 
   return (
     <Page>
-      {/* Page header */}
-      <PageSection variant={PageSectionVariants.light}>
-        <Flex
-          spaceItems={{ default: 'spaceItemsMd' }}
-          alignItems={{ default: 'alignItemsFlexStart' }}
-          flexWrap={{ default: 'nowrap' }}
-        >
-          <FlexItem>
-            <Label color="blue">Account</Label>
-          </FlexItem>
-          <FlexItem>
-            <Title headingLevel="h1" size="2xl">
-              {accountName}
-            </Title>
-          </FlexItem>
-        </Flex>
-        {/* Page tabs */}
-      </PageSection>
-      <PageSection type="tabs" variant={PageSectionVariants.light}>
-        <Tabs activeKey={activeTabKey} onSelect={handleTabClick} usePageInsets id="open-tabs-example-tabs-list">
-          <Tab eventKey={0} title={<TabTitleText>Details</TabTitleText>} tabContentId={`tabContent${0}`} />
-          <Tab eventKey={1} title={<TabTitleText>Clusters</TabTitleText>} tabContentId={`tabContent${1}`} />
-        </Tabs>
-      </PageSection>
-      <PageSection variant={PageSectionVariants.light}>
-        <TabContent key={0} eventKey={0} id={`tabContent${0}`} activeKey={activeTabKey} hidden={0 !== activeTabKey}>
-          <TabContentBody>{detailsTabContent}</TabContentBody>
-        </TabContent>
-        <TabContent key={1} eventKey={1} id={`tabContent${1}`} activeKey={activeTabKey} hidden={1 !== activeTabKey}>
-          <TabContentBody>{clustersTabContent}</TabContentBody>
-        </TabContent>
-      </PageSection>
+      <AccountsHeader accountName={accountName} label="Account" />
+      <AccountsTabs
+        detailsTabContent={<AccountDetailsContent loading={loading} accountData={accountData} />}
+        clustersTabContent={<AccountClusters />}
+      />
     </Page>
   );
 };
