+++ src/app/utils/debugLogs.ts
@@ -1,0 +1,5 @@
+export const debug = (...args: unknown[]) => {
+  if (import.meta.env.DEV) {
+    console.log(...args);
+  }
+};
